backend/pyproject.toml
[project]
name = "sql-mcp"
version = "0.1.0"
description = "BigQuery MCP Server with Google ADK integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.14.1",
    "mcp>=1.1.2",
    "google-cloud-bigquery>=3.11.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "toolbox-core>=0.5.2",
    "langgraph>=0.6.7",
    "langchain-core>=0.3.76",
    "langchain-google-vertexai>=2.1.2",
    "langchain[google-genai]>=0.3.27",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
]
